put auxiliary libs into a separate workflow
