+++ .github/workflows/test-changed-auxiliary-libs.yml
@@ -1,10 +1,10 @@
-name: Test Modified Packages
+name: Test rxFire, @firebase/testing and @firebase/rules-unit-testing
 
 on: pull_request
 
 jobs:
   test:
-    name: Test Packages With Changed Files
+    name: Test rxFire, @firebase/testing and @firebase/rules-unit-testing
     runs-on: ubuntu-latest
 
     steps:
@@ -25,5 +25,5 @@ jobs:
       run: |
         cp config/ci.config.json config/project.json
         yarn
-    - name: Run tests on changed packages
-      run: xvfb-run yarn test:changed
+    - name: Run tests
+      run: xvfb-run yarn test:changed:auxiliary
