+++ package.json
@@ -38,9 +38,10 @@
     "pretest:coverage": "mkdirp coverage",
     "ci:coverage": "lcov-result-merger 'packages/**/lcov.info' 'lcov-all.info'",
     "test:coverage": "lcov-result-merger 'packages/**/lcov.info' | coveralls",
-    "test:changed": "ts-node-script scripts/ci-test/run_changed_no_firestore.ts",
+    "test:changed": "ts-node-script scripts/ci-test/run_changed_core.ts",
     "test:changed:firestore": "ts-node-script scripts/ci-test/run_changed_firestore.ts",
     "test:changed:fcm": "ts-node-script scripts/ci-test/run_changed_fcm_integration.ts",
+    "test:changed:auxiliary": "ts-node-script scripts/ci-test/run_changed_auxiliary_libs.ts",
     "test:setup": "node tools/config.js",
     "test:saucelabs": "node scripts/run_saucelabs.js",
     "docgen:js": "node scripts/docgen/generate-docs.js --api js",
