+++ scripts/ci-test/run_changed_core.ts
@@ -1,0 +1,41 @@
+/**
+ * @license
+ * Copyright 2020 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import { runTests, getTestTasks, createTestTask } from './run_changed';
+import { buildForTests } from './build';
+
+const ignoredPackages = [
+  '@firebase/firestore',
+  'firebase-firestore-integration-test',
+  'firebase-messaging-integration-test',
+  'firebase-namespace-integration-test',
+  '@firebase/testing',
+  '@firebase/rules-unit-testing',
+  'rxfire'
+];
+
+async function run() {
+  let testTasks = await getTestTasks();
+
+  // remove the ignored packages from the tasks
+  testTasks = testTasks.filter(t => !ignoredPackages.includes(t.pkgName));
+  await buildForTests(testTasks);
+
+  runTests(testTasks);
+}
+
+run();
